Refactor passes into a submodule
