+++ jeff65/gold/passes/asm.py
@@ -15,7 +15,7 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import struct
-from . import ast, storage
+from .. import ast, storage
 
 
 class AssemblyError(Exception):
