+++ jeff65/gold/passes/binding.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from . import ast
+from .. import ast
 
 
 class ScopedPass(ast.TranslationPass):
